Struct ethcore_transaction::Transaction

pub struct Transaction {
    pub nonce: U256,
    pub gas_price: U256,
    pub gas: U256,
    pub action: Action,
    pub value: U256,
    pub data: Vec<u8>,
}

A set of information describing an externally-originating message call or contract creation operation.

Fields
nonce: U256

Nonce.

gas_price: U256

Gas price.

gas: U256

Gas paid up front for transaction execution.

action: Action

Action, can be either call or contract create.

value: U256

Transfered value.

data: Vec<u8>

Transaction data.

Methods

impl Transaction

pub fn rlp_append_unsigned_transaction(
&self,
s: &mut RlpStream,
chain_id: Option<u64>
)

Append object with a without signature into RLP stream

impl Transaction

pub fn hash(&self, chain_id: Option<u64>) -> H256

The message hash of the transaction.

pub fn sign(self, secret: &Secret, chain_id: Option<u64>) -> SignedTransaction

Signs the transaction as coming from sender.

pub fn with_signature(
self,
sig: Signature,
chain_id: Option<u64>
) -> UnverifiedTransaction

Signs the transaction with signature.

pub fn fake_sign(self, from: Address) -> SignedTransaction

Specify the sender; this won't survive the serialize/deserialize process, but can be cloned.

pub fn null_sign(self, chain_id: u64) -> SignedTransaction

Add EIP-86 compatible empty signature.

pub fn gas_required_for(
is_create: bool,
data: &[u8],
schedule: &Schedule
) -> u64

Get the transaction cost in gas for the given params.

pub fn gas_required(&self, schedule: &Schedule) -> u64

Get the transaction cost in gas for this transaction.